requirements.txt
albumentations==1.4.6
detectron2==0.6
einops==0.8.1
fvcore==0.1.5.post20221221
matplotlib==3.8.4
monai==0.9.0
mpi4py==4.0.3
numpy==2.2.3
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==2.2.3
Pillow==11.1.0
PyYAML==6.0.1
PyYAML==6.0.2
scipy==1.15.2
setuptools==69.5.1
SimpleITK==2.3.1
SimpleITK==2.4.1
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.4
